Move mempool.space calls into a separate module
